update cube creation with scryfall api calls
previously cube was just a collection of card names pulled from cube
cobra, this commit now calls upon the scryfall api to fetch the card
image, cmc and rarity upon initial cube creation.

the scryfall card objects have a lot of information in them, most of it
is ignored currently but models can easily be extended in future.